schema: fix semantics for names in namespaces
The spec doesn't make a distinction between the namespace
field and a namespace implied by a full name, so we change
to fix that.

Also, when we ask a record for its schema, we always want to
include the full name because there's no parent
to inherit the namespace from, so fix that too.